Refactor joinWithSeparator method to JsonIssueParser and remove duplicates from DetectifyParser and SnykParser
